+++ src/plone/mls/listing/api.py
@@ -7,14 +7,24 @@ import logging
 import time
 
 # zope imports
+from Acquisition import aq_parent, aq_inner
+from Products.CMFPlone.interfaces import IPloneSiteRoot
+from plone import api
 from plone.registry.interfaces import IRegistry
+from zope.annotation.interfaces import IAnnotations
 from zope.component import getUtility
+
 
 # local imports
 from mls.apiclient.client import ListingResource
 from mls.apiclient.exceptions import MLSError
-from plone.mls.core.interfaces import IMLSSettings
+from plone.mls.core.api import get_settings
 from plone.mls.listing import PRODUCT_NAME
+from plone.mls.listing.browser.localconfig import CONFIGURATION_KEY
+from plone.mls.listing.interfaces import (
+    ILocalAgencyInfo,
+    IMLSAgencyContactInformation,
+)
 
 
 logger = logging.getLogger(PRODUCT_NAME)
@@ -61,10 +71,11 @@ class SearchOptions(object):
     language = None
     timeout = 100
 
-    def __init__(self, category, language, timeout):
+    def __init__(self, category, language, timeout, context=None):
         self.category = category
         self.language = language
         self.timeout = timeout
+        self.context = context
 
         self._items = []
         self._loaded = False  # Is the category already loaded?
@@ -123,10 +134,9 @@ class SearchOptions(object):
         if self.category is not None:
             self._last_update_time_in_minutes = time.time() / 60
             self._last_update_time = DateTime()
-            registry = getUtility(IRegistry)
-            settings = registry.forInterface(IMLSSettings)
-            base_url = getattr(settings, 'mls_site', None)
-            api_key = getattr(settings, 'mls_key', None)
+            settings = get_settings(context=self.context)
+            base_url = settings.get('mls_site', None)
+            api_key = settings.get('mls_key', None)
             resource = ListingResource(base_url, api_key=api_key)
             results = []
             try:
@@ -149,7 +159,7 @@ class SearchOptions(object):
         return self._items
 
 
-def search_options(mls_url, category, lang=None):
+def search_options(mls_url, category, lang=None, context=None):
     if mls_url is None or len(mls_url) < 1:
         return
 
@@ -158,7 +168,7 @@ def search_options(mls_url, category, lang=None):
     options = OPTIONS_CACHE.get(key, None)
 
     if options is None:
-        options = SearchOptions(category, lang, timeout)
+        options = SearchOptions(category, lang, timeout, context=context)
 
         if options.update():
             OPTIONS_CACHE[key] = options
@@ -168,17 +178,16 @@ def search_options(mls_url, category, lang=None):
     return options.items
 
 
-def recent_listings(params={}, batching=True):
+def recent_listings(params={}, batching=True, context=None):
     """Return a list of recent MLS listings."""
     search_params = {
         'sort_on': 'created',
         'reverse': '1',
     }
     search_params.update(params)
-    registry = getUtility(IRegistry)
-    settings = registry.forInterface(IMLSSettings)
-    base_url = getattr(settings, 'mls_site', None)
-    api_key = getattr(settings, 'mls_key', None)
+    settings = get_settings(context=context)
+    base_url = settings.get('mls_site', None)
+    api_key = settings.get('mls_key', None)
     batch = None
     results = []
     resource = ListingResource(base_url, api_key=api_key)
@@ -193,12 +202,11 @@ def recent_listings(params={}, batching=True):
     return results
 
 
-def listing_details(listing_id, lang=None):
+def listing_details(listing_id, lang=None, context=None):
     """Return detail information for a listing."""
-    registry = getUtility(IRegistry)
-    settings = registry.forInterface(IMLSSettings)
-    base_url = getattr(settings, 'mls_site', None)
-    api_key = getattr(settings, 'mls_key', None)
+    settings = get_settings(context=context)
+    base_url = settings.get('mls_site', None)
+    api_key = settings.get('mls_key', None)
     resource = ListingResource(base_url, api_key=api_key)
     try:
         listing = resource.get(listing_id, lang=lang)
@@ -208,17 +216,16 @@ def listing_details(listing_id, lang=None):
     return listing.get('listing', None)
 
 
-def search(params={}, batching=True):
+def search(params={}, batching=True, context=None):
     """Search for listings."""
     search_params = {
         'sort_on': 'created',
         'reverse': '1',
     }
     search_params.update(params)
-    registry = getUtility(IRegistry)
-    settings = registry.forInterface(IMLSSettings)
-    base_url = getattr(settings, 'mls_site', None)
-    api_key = getattr(settings, 'mls_key', None)
+    settings = get_settings(context=context)
+    base_url = settings.get('mls_site', None)
+    api_key = settings.get('mls_key', None)
     batch = None
     results = []
     resource = ListingResource(base_url, api_key=api_key)
@@ -231,3 +238,52 @@ def search(params={}, batching=True):
     if batching:
         return results, batch
     return results
+
+
+def _local_agency_info(context):
+    """Get local agency information."""
+    settings = None
+    obj = context
+    while (
+            not IPloneSiteRoot.providedBy(obj) and
+            not ILocalAgencyInfo.providedBy(obj)):
+        parent = aq_parent(aq_inner(obj))
+        if parent is None:
+            return
+        obj = parent
+    if ILocalAgencyInfo.providedBy(obj):
+        annotations = IAnnotations(obj)
+        settings = annotations.get(
+            CONFIGURATION_KEY, annotations.setdefault(CONFIGURATION_KEY, {}))
+    return settings
+
+
+def get_agency_info(context=None):
+    """Get the agency information."""
+    if context is None:
+        context = api.portal.get()
+
+    local_info = _local_agency_info(context)
+    if local_info is not None:
+        logger.debug('Returning local agency information.')
+        return local_info
+
+    # Get the global agency info.
+    settings = {}
+    registry = getUtility(IRegistry)
+    if registry is not None:
+        try:
+            registry_settings = registry.forInterface(
+                IMLSAgencyContactInformation
+            )
+        except:
+            logger.warning('Global agency information not available.')
+        else:
+            settings = dict([
+                (a, getattr(registry_settings, a)) for a in
+                registry_settings.__schema__]
+            )
+            logger.debug('Returning global agency information.')
+    if not settings.get('use_custom_info', False):
+        return
+    return settings